DRYing and cleaning
successfully outsourced a few types to a shared injector

create injectorSOR

get rid of unused dependencies

DRYing handlers
