+++ lib/handlers/handler.ts
@@ -282,10 +282,6 @@ export abstract class APIGLambdaHandler<
   > {
     let bodyRaw: any;
 
-    log.info('event information', {
-      event
-    })
-
     if (event.body) {
       try {
         bodyRaw = JSON.parse(event.body);
@@ -306,11 +302,6 @@ export abstract class APIGLambdaHandler<
     let queryParamsRaw: APIGatewayProxyEventQueryStringParameters | null =
       event.queryStringParameters;
     const queryParamsSchema = this.requestQueryParamsSchema();
-
-    log.info('queryParamsSchema', {
-      queryParamsSchema
-    })
-
 
     let queryParams: ReqQueryParams | undefined;
     if (queryParamsRaw && queryParamsSchema) {
