--- test.json
@@ -1,42 +1,0 @@
-{
-    "blockNumber": "13407481",
-    "amount": "100",
-    "amountDecimals": "0.0000000000000001",
-    "quote": "14863",
-    "quoteDecimals": "0.000000000000014863",
-    "quoteGasAdjusted": "-1161910637082771893",
-    "quoteGasAdjustedDecimals": "-1.161910637082771893",
-    "gasUseEstimateQuote": "1161910637082786756",
-    "gasUseEstimateQuoteDecimals": "1.161910637082786756",
-    "gasUseEstimate": "113000",
-    "gasUseEstimateUSD": "27.113563",
-    "gasPriceWei": "68486564633",
-    "route": [
-        [
-            {
-                "type": "v3-pool",
-                "address": "0x360b9726186C0F62cc719450685ce70280774Dc8",
-                "tokenIn": {
-                    "chainId": 1,
-                    "decimals": "18",
-                    "address": "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2",
-                    "symbol": "WETH"
-                },
-                "tokenOut": {
-                    "chainId": 1,
-                    "decimals": "18",
-                    "address": "0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984",
-                    "symbol": "UNI"
-                },
-                "fee": "10000",
-                "liquidity": "7529515858212782178449",
-                "sqrtRatioX96": "6465991549558817900743090467",
-                "tickCurrent": "-50119",
-                "amountIn": "100",
-                "amountOut": "14863"
-            }
-        ]
-    ],
-    "routeString": "100.00% = WETH -- 1% --> UNI",
-    "quoteId": "8f656"
-}
